requirements.txt
pillow
tensorflow>=2.15.0
numpy
pandas
pillow
huggingface_hub
matplotlib
pytest